feat(terminal-types): secrets plugin — per-project tab for secrets config
Closes bd remote-dev-1ebu.3.
