+++ src/lib/terminal-plugins/init-server.ts
@@ -22,6 +22,7 @@ import { PortManagerServerPlugin } from "./plugins/port-manager-plugin-server";
 import { TrashServerPlugin } from "./plugins/trash-plugin-server";
 import { ProjectPrefsServerPlugin } from "./plugins/project-prefs-plugin-server";
 import { GitHubMaintenanceServerPlugin } from "./plugins/github-maintenance-plugin-server";
+import { SecretsServerPlugin } from "./plugins/secrets-plugin-server";
 import { createLogger } from "@/lib/logger";
 
 const log = createLogger("PluginInit.Server");
@@ -54,6 +55,7 @@ export function initializeServerPlugins(): void {
   TerminalTypeServerRegistry.register(TrashServerPlugin, { builtIn: true });
   TerminalTypeServerRegistry.register(ProjectPrefsServerPlugin, { builtIn: true });
   TerminalTypeServerRegistry.register(GitHubMaintenanceServerPlugin, { builtIn: true });
+  TerminalTypeServerRegistry.register(SecretsServerPlugin, { builtIn: true });
 
   TerminalTypeServerRegistry.setDefaultType("shell");
